fix: refactor backend for vercel serverless deployment
