+++ api/index.js
@@ -1,0 +1,3 @@
+const app = require('../backend/src/app');
+
+module.exports = app;
