+++ backend/src/app.js
@@ -1,0 +1,56 @@
+require('./config/loadEnv');
+
+process.on('uncaughtException', (err) => {
+    console.error('Uncaught Exception:', err);
+});
+process.on('unhandledRejection', (reason, promise) => {
+    console.error('Unhandled Rejection at:', promise, 'reason:', reason);
+});
+
+const express = require('express');
+const cors = require('cors');
+
+const app = express();
+
+// Use permissive CORS to ensure E2E works locally
+app.use(cors());
+app.use(express.json());
+
+// Request logging for debugging
+app.use((req, res, next) => {
+    console.log('Incoming request:', req.method, req.url);
+    next();
+});
+
+const swaggerUi = require('swagger-ui-express');
+const swaggerSpecs = require('./config/swagger');
+
+const authRoutes = require('./routes/auth');
+const taskRoutes = require('./routes/tasks');
+const timeEntryRoutes = require('./routes/timeEntries');
+const uploadRoutes = require('./routes/upload');
+const profileRoutes = require('./routes/profile');
+const aiRoutes = require('./routes/ai');
+
+app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpecs));
+app.use('/api/auth', authRoutes);
+
+app.use('/api/tasks', taskRoutes);
+app.use('/api/time-entries', timeEntryRoutes);
+app.use('/api/upload', uploadRoutes);
+app.use('/api/profile', profileRoutes);
+app.use('/api/ai', aiRoutes);
+
+app.get('/', (req, res) => {
+    res.send('Task Manager Backend is running!');
+});
+
+app.get('/health', (req, res) => {
+    res.status(200).json({
+        status: 'healthy',
+        timestamp: new Date().toISOString(),
+        environment: process.env.NODE_ENV || 'development'
+    });
+});
+
+module.exports = app;
